requirements.txt
pandas
numpy
matplotlib
scikit-learn
jupyter
sqlalchemy
psycopg2
openpyxl
selenium
python-dotenv
